[REFACTOR #35 ] jwtProvider Refactoring
