--- pass-service/src/main/java/com/farmted/passservice/service/TokenService.java
@@ -1,10 +1,0 @@
-package com.farmted.passservice.service;
-
-import com.farmted.passservice.enums.RoleEnums;
-import com.farmted.passservice.enums.TokenType;
-
-
-public interface TokenService {
-    String createToken(String uuid, RoleEnums role, TokenType tokenType);
-    void saveRefreshToken(String uuid, String refreshToken);
-}
