+++ pass-service/src/main/java/com/farmted/passservice/util/jwt/JwtFilter.java
@@ -41,6 +41,8 @@ public class JwtFilter extends OncePerRequestFilter {
                 // AccessToken과 RefreshToken을 쿠키에서 추출
                 if (cookie.getName().equals(JwtProvider.AUTH_HEADER)) {
                     accessToken = jwtProvider.resolveToken(cookie);
+                } else if (cookie.getName().equals(JwtProvider.REFRESH_HEADER)) {
+                    refreshToken = jwtProvider.resolveToken(cookie);
                 }
             }
         }
@@ -70,6 +72,11 @@ public class JwtFilter extends OncePerRequestFilter {
                     setAuthentication(uuid);
                 }
             }
+            // accessToken이 null 인데 refreshToken이 있을 경우
+        } else if (refreshToken != null) {
+            if (jwtProvider.validateRefreshToken(refreshToken)) {
+                setAuthentication(jwtProvider.getUserInfoFromToken(refreshToken).getSubject());
+            }
         }
         filterChain.doFilter(request, response);
     }
@@ -78,10 +85,14 @@ public class JwtFilter extends OncePerRequestFilter {
     public void setAuthentication(String uuid) {
         // 빈 SecurityContext 생성
         SecurityContext context = SecurityContextHolder.createEmptyContext();
-        // JwtProvider의 createAuth를 사용하여 Authentication 객체 생성
+        /*
+        * JwtProvider의 createAuth를 사용하여 Authentication 객체 생성
+        * 생성된 Authentication 객체를 SecurityContext에 설정 후
+        * SecurityContextHolder를 새로운 컨텍스트로 업데이트.
+        *   -> 사용자를 인증된 사용자로 인식할 수 있음
+        * */
         Authentication authentication = jwtProvider.createAuth(uuid);
         context.setAuthentication(authentication);
-
         SecurityContextHolder.setContext(context);
     }
 }
