+++ pass-service/src/main/java/com/farmted/passservice/util/oauth2/handler/OAuth2LoginSuccessHandler.java
@@ -2,7 +2,6 @@ package com.farmted.passservice.util.oauth2.handler;
 
 import com.farmted.passservice.enums.RoleEnums;
 import com.farmted.passservice.enums.TokenType;
-import com.farmted.passservice.service.TokenService;
 import com.farmted.passservice.util.jwt.JwtProvider;
 import com.farmted.passservice.util.oauth2.CustomOAuth2User;
 import jakarta.servlet.ServletException;
@@ -22,7 +21,6 @@ import java.io.IOException;
 public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
 
     private final JwtProvider jwtProvider;
-    private final TokenService tokenService;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
@@ -36,12 +34,12 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
             // Pass의 Role이 GUEST일 경우 회원가입이 완료되지 않은 회원이므로 회원 상세정보 적기 페이지로 리다이렉트
             if (role == RoleEnums.GUEST) {
                 jwtProvider.setToken(token, TokenType.ACCESS, response);
-                String refreshToken = tokenService.createToken(uuid, role, TokenType.REFRESH);
-                tokenService.saveRefreshToken(uuid, refreshToken);
+                String refreshToken = jwtProvider.createToken(uuid, role, TokenType.REFRESH);
+                jwtProvider.saveRefreshToken(uuid, refreshToken);
                 response.sendRedirect("/home");
             } else {
                 jwtProvider.setToken(token, TokenType.ACCESS, response);
-                tokenService.saveRefreshToken(uuid, token);
+                jwtProvider.saveRefreshToken(uuid, token);
                 response.sendRedirect("/home");
             }
         } catch (Exception e) {
